refactor: 전략패턴 적용 docs: 문서화 적용
