+++ src/main/java/step3/README.md
@@ -45,3 +45,38 @@ chore (maintain)
 2. 자동차가 전진하는 조건을 구현한다
 3. 자동차가 이동하는 기능을 구현한다
 4. 자동차의 상태를 화면에 출력하는 기능을 구현한다
+
+## 구현 설명
+
+### util
+
+`NumberValidator`
+
+- 입력받은 숫자가 각각 자동차의 수와 이동 횟수에 적합한 숫자인지 확인하는 유틸성 클래스
+
+`RandomNumberGenerator`
+
+- 0 ~ 9 사이의 랜덤한 숫자를 생성하는 유틸성 클래스
+
+`RacingCreateHelper`
+
+- 여러개의 객체 생성을 도와주는 유틸성 클래스
+
+### domain
+
+`Racecource`
+- 자동차 경주를 진행하는 레이스코스를 나타내는 클래스
+- 자동차를 입장시키면 비어있는 트랙에 자동차를 입장시킨다
+- 이동신호를 주면 모든 트랙에 있는 자동차에게 이동신호를 준다
+
+`Track`
+- 자동차 경주를 진행하는 트랙을 나타내는 클래스
+- 트랙 내부 자동차의 위치를 나타내는 `CarPosition`을 가진다
+- 전진신호를 자동차에게 전달할 수 있다
+
+`RacingCar`
+- 자동차 경주를 진행하는 자동차를 나타내는 클래스
+- 전진 시도시 랜덤한 숫자를 생성하여 전진 여부를 판단한다
+- 전진 판단시 1만큼 전진한다
+- 전진여부를 판단하는 기준은 `RandomNumberGenerator`를 통해 생성된 랜덤한 숫자가 4이상인 경우이다
+- 기본 전진 전략으로 `RandomMove` 전략을 사용하며, 확장성을 위해 `MoveStrategy` 인터페이스를 구현한다
