+++ src/main/java/step3/domain/Racecourse.java
@@ -5,7 +5,7 @@ import java.util.List;
 import java.util.Optional;
 
 public class Racecourse {
-    private List<Track> tracks = new ArrayList<>();
+    private final List<Track> tracks = new ArrayList<>();
 
     public void addTrack(Track track) {
         tracks.add(track);
