+++ src/main/java/step3/move/MoveStrategy.java
@@ -1,0 +1,5 @@
+package step3.move;
+
+public interface MoveStrategy {
+    int moveAmount();
+}
