+++ src/test/java/step3/component/NumberValidatorTest.java
@@ -3,6 +3,7 @@ package step3.component;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step3.util.NumberValidator;
 
 class NumberValidatorTest {
